% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_utility.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Convert a \code{Map} Object to a Data Frame}
\usage{
as.data.frame(x)
}
\arguments{
\item{x}{An input \code{Map} object.}
}
\value{
A data frame having the map informtion.
}
\description{
This inherited S3 function is devised to convert a Map object to a data frame. The Map
object has six slots: Marker name (Name), Chromosome, Positin, minor alllel frequency (Maf), Tyep,
and Status.
}
\examples{
data(bov80K)
class(bov80K)
tmpdat<-as.data.frame(bov80K)
class(tmpdat)

}
\author{
Nick X-L Wu
}
