% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_class.R
\docType{data}
\name{bov80K}
\alias{bov80K}
\title{A \code{Map} object with 80K bovine SNPs}
\format{A Map object with a list of 30 Chrom objects.}
\usage{
data(bov80K)
}
\description{
This example \code{Map} object contains map information for 76,694 SNPs on 30 chromosomes, with chromosome X
represnted by 30. Each \code{Chrom} object has six slots: Locus names (Name), Chromosome, Position, minor
allele frquency (Maf), Type and Status.
}
\keyword{datasets}
