% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_class.R
\docType{class}
\name{Chrom}
\alias{Chrom}
\alias{Chrom-class}
\title{The Chromosome class}
\description{
This is a S4 \code{Chrom} classs for a single chromosome
}
\details{
This \code{Chrom} class abstracts the information for all the markers on a single chromosome. It has
the following slots.
}
\section{Slots}{

 \describe{
   \item{\code{Chromosome}:}{A character string as the chromosome name or index.}
   \item{\code{Name}:}{A character vector containing the names for all the markers on this chromosome.}
   \item{\code{Position}:}{A numeric vector containing the positions of all the markers on this chromosome.}
   \item{\code{Maf}:}{A numeric vector containing the minor allele frequencies of all the markers on
    this chromosome.}
   \item{\code{Type}:}{A character vector containing the types for all the markers on this chromosome.}
   \item{\code{Status}:}{An integer vector containing the statuses for all the markers on this chromosome.}
 }
}

\author{
Nick X-L Wu
}
