% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_class.R
\docType{class}
\name{Map}
\alias{Map}
\alias{Map-class}
\title{The Map class}
\description{
This is a S4 \code{Map} class representing a genome.
}
\details{
This \code{Map} class has information of all the markers on a map, which represents either entirely
or partially a genome. It consists of multiple chromosomes, though a \code{Map} object can have only one
chromosome.
}
\section{Slots}{

 \describe{
   \item{\code{.Data}:}{A list of \code{Chrom} objects.}
 }
}

\author{
Nick X-L Wu
}
