% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_utility.R
\docType{methods}
\name{namesOfChrom}
\alias{namesOfChrom}
\alias{namesOfChrom,Chrom-method}
\alias{namesOfChrom,Chrom}
\alias{namesOfChrom,Map-method}
\alias{namesOfChrom,Map}
\title{Extract chromosome indexes or names}
\usage{
namesOfChrom(object)

\S4method{namesOfChrom}{Chrom}(object)

\S4method{namesOfChrom}{Map}(object)
}
\arguments{
\item{object}{An input object, which can be a \code{Chrom} object, a \code{Map} object, or a data frame.}
}
\value{
Chromosome names of the input object.
}
\description{
This S4 function is used to extract chromosome indexes or names for \code{Chrom} object,
a \code{Map} object, or a data frame.
}
\examples{
data("bov80K")
namesOfChrom(bov80K)

}
\author{
Nick X-L Wu
}
