% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_utility.R
\docType{methods}
\name{numberOfLoci}
\alias{numberOfLoci}
\alias{numberOfLoci,Chrom-method}
\alias{numberOfLoci,Chrom}
\alias{numberOfLoci,Map-method}
\alias{numberOfLoci,Map}
\title{Extract the number of loci on each chromosome}
\usage{
numberOfLoci(object)

\S4method{numberOfLoci}{Chrom}(object)

\S4method{numberOfLoci}{Map}(object)
}
\arguments{
\item{object}{An input object, which can be a \code{Chrom} object, a \code{Map} object, or a data frame.}
}
\value{
Number of SNP loci by chrommosomes.
}
\description{
This S4 function is used to extract the number of loci on each chromsome.
}
\examples{
data("bov80K")
numberOfLoci(bov80K)

}
\author{
Nick X-L Wu
}
