% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_utility.R
\docType{methods}
\name{plotMap}
\alias{plotMap}
\alias{plotMap,Map-method}
\title{Plotting a Map object.}
\usage{
plotMap(object, ...)

\S4method{plotMap}{Map}(object, plotMaf = FALSE, bin = 1e+06, maf.base = 0.3)
}
\arguments{
\item{object}{An input Map object.}

\item{...}{Some extra input parameter values; optional.}

\item{plotMaf}{A logical value. If TRUE, MAF density will be plotted; otherwise, SNPs are plotted 
based on their map positions. 
plotted based on their chromosomal po}

\item{bin}{Length of the bin in which SNPs are counted. The default length is 1M bp.}

\item{maf.base}{The baseline MAF, which by default is 0.3. A MAF density above maf.base
is trucated to be 1.}
}
\value{
None.
}
\description{
This S4 generic function is used to plot a Map object. 
The Map object needs to have at leaset three slots: Marker names (Name), 
Chromosome, and Position.
}
\author{
Nick X-L Wu
}
