% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSNPs_utility.R
\name{summary}
\alias{summary}
\title{Summary statiscs for a Map object.}
\usage{
summary(x)
}
\arguments{
\item{x}{An input map object.}
}
\value{
Summary statistics of the \code{Map} object.
}
\description{
This inherited S3 function is devised specifically for compute summary statistics for
a Map object. The output includes Chromosome names or indexes (chrom), numuber of SNPs on each
chromosome (nLoci), minimum map position on each chromosome (Min), maximum map position on each
chromosome (Max), chromosome length (Length), average spacing or bin width (mu.bw), standard
deviation of average spacing (sd.bw), minimum spacing (min.bw), and maximum spacing or gap
length (max.bw).
}
\examples{
data("bov80K")
summary(bov80K)

}
\author{
Nick X-L Wu
}
